split training and validation data by tracks not slices
The chopping and normalization is done before the training,
when retrieving the training and validation set.

In the h5 file the unmodified spectra are saved.
The spectra do not have the same shape and cannot be saved all at once.
Therefore put them in subgroups and save them track by track.

Calculate mashup, instrumental and acapella spectra for all tracks.
The data can be used to train both, so they have to be calculated only once.